usermodel2/requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
beautifulsoup4==4.9.0
cffi==1.14.0
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.1.0
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
Faker==4.0.3
lazy-object-proxy==1.4.3
misaka==2.1.1
Pillow==7.1.2
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
wrapt==1.12.1
